+++ packages/genesys/packages/react-components-core/src/Dialog/index.ts
@@ -1,0 +1,3 @@
+export * from "./Dialog.types";
+export * from "./hooks";
+export * from "./Dialog.constants";