+++ packages/genesys/packages/react-components-core/src/config/components.config.types.ts
@@ -1,5 +1,12 @@
 import { BlockchainLinks, ComponentConfig } from "./config.types";
+import { ButtonVariant } from "../Button";
 
 export type ExtendedCoreBlockchainAddressConfig<P> = ComponentConfig<P> & {
     blockchainLinks: BlockchainLinks;
+};
+
+export type ExtendedCoreDialogConfig<P> = ComponentConfig<P> & {
+    actions: {
+        variant: ButtonVariant;
+    };
 };