+++ packages/genesys/packages/react-components-core/src/index.tsx
@@ -4,6 +4,7 @@ export * from "./Alert";
 export * from "./Button";
 export * from "./Checkbox";
 export * from "./Chip";
+export * from "./Dialog";
 export * from "./Form";
 export * from "./FormControl";
 export * from "./FormControlError";