+++ packages/genesys/packages/react-components/src/Dialog/hooks/useDialog.ts
@@ -1,8 +1,5 @@
-import {
-    DialogProps,
-    useDialog,
-    UseDialogResult as CoreUseDialogResult,
-} from "@peersyst/react-components-core";
+import { useDialog, UseDialogResult as CoreUseDialogResult } from "@peersyst/react-components-core";
+import { DialogProps } from "../Dialog.types";
 import Dialog from "../Dialog";
 
 export interface UseDialogResult extends Omit<CoreUseDialogResult, "showDialog"> {