+++ packages/genesys/packages/react-components/src/Dialog/index.tsx
@@ -1,4 +1,4 @@
 export { default as Dialog } from "./Dialog";
+export * from "./Dialog.types";
 export * from "./Dialog.styles";
 export * from "./hooks";
-export type { DialogProps, DialogButton, DialogOptionProps } from "@peersyst/react-components-core";