+++ packages/genesys/packages/react-native-components/src/config/components.config.tsx
@@ -62,6 +62,9 @@ const componentsConfig: ComponentsConfig = {
     },
     Dialog: {
         defaultProps: {},
+        actions: {
+            variant: "text",
+        },
     },
     Modal: {
         defaultProps: {